Move statefulsets code to seperate files;
