+++ caas/kubernetes/provider/k8s.go
@@ -1134,7 +1134,7 @@ func (k *kubernetesClient) EnsureService(
 		cleanups = append(cleanups, cleanUpDaemonSet)
 	default:
 		// This should never happend because we have validated both in this method and in `charm.v6`.
-		return errors.NotSupportedf("deployment type %q not supported", params.Deployment.DeploymentType)
+		return errors.NotSupportedf("deployment type %q", params.Deployment.DeploymentType)
 	}
 	return nil
 }
@@ -1572,142 +1572,6 @@ func getPodManagementPolicy(svc *specs.ServiceSpec) (out apps.PodManagementPolic
 		// no need to consider other cases because we have done validation in podspec parsing stage.
 	}
 	return out
-}
-
-func (k *kubernetesClient) configureStatefulSet(
-	appName, deploymentName string, annotations k8sannotations.Annotation, workloadSpec *workloadSpec,
-	containers []specs.ContainerSpec, replicas *int32, filesystems []storage.KubernetesFilesystemParams,
-) error {
-	logger.Debugf("creating/updating stateful set for %s", appName)
-
-	// Add the specified file to the pod spec.
-	cfgName := func(fileSetName string) string {
-		return applicationConfigMapName(deploymentName, fileSetName)
-	}
-
-	existingStatefulSet, err := k.getStatefulSet(deploymentName)
-	if err != nil && !errors.IsNotFound(err) {
-		return errors.Trace(err)
-	}
-	existing := err == nil
-
-	var randPrefix string
-	// Include a random snippet in the pvc name so that if the same app
-	// is deleted and redeployed again, the pvc retains a unique name.
-	// Only generate it once, and record it on the stateful set.
-	if existing {
-		randPrefix = existingStatefulSet.Annotations[labelApplicationUUID]
-	}
-	if randPrefix == "" {
-		randPrefix, err = k.randomPrefix()
-		if err != nil {
-			return errors.Trace(err)
-		}
-	}
-
-	statefulset := &apps.StatefulSet{
-		ObjectMeta: v1.ObjectMeta{
-			Name: deploymentName,
-			Annotations: k8sannotations.New(nil).
-				Merge(annotations).
-				Add(labelApplicationUUID, randPrefix).ToMap(),
-		},
-		Spec: apps.StatefulSetSpec{
-			Replicas: replicas,
-			Selector: &v1.LabelSelector{
-				MatchLabels: map[string]string{labelApplication: appName},
-			},
-			Template: core.PodTemplateSpec{
-				ObjectMeta: v1.ObjectMeta{
-					Labels:      map[string]string{labelApplication: appName},
-					Annotations: podAnnotations(annotations.Copy()).ToMap(),
-				},
-			},
-			PodManagementPolicy: getPodManagementPolicy(workloadSpec.Service),
-			ServiceName:         headlessServiceName(deploymentName),
-		},
-	}
-	podSpec := workloadSpec.Pod
-	if err := k.configurePodFiles(appName, annotations, &podSpec, containers, cfgName); err != nil {
-		return errors.Trace(err)
-	}
-	existingPodSpec := podSpec
-
-	// Create a new stateful set with the necessary storage config.
-	legacy := isLegacyName(deploymentName)
-	if err := k.configureStorage(&podSpec, &statefulset.Spec, appName, randPrefix, legacy, filesystems); err != nil {
-		return errors.Annotatef(err, "configuring storage for %s", appName)
-	}
-	statefulset.Spec.Template.Spec = podSpec
-	return k.ensureStatefulSet(statefulset, existingPodSpec)
-}
-
-func (k *kubernetesClient) ensureStatefulSet(spec *apps.StatefulSet, existingPodSpec core.PodSpec) error {
-	_, err := k.createStatefulSet(spec)
-	if errors.IsNotValid(err) {
-		return errors.Annotatef(err, "ensuring stateful set %q", spec.GetName())
-	}
-	if err != nil && !errors.IsAlreadyExists(err) {
-		return errors.Trace(err)
-	}
-	// The statefulset already exists so all we are allowed to update is replicas,
-	// template, update strategy. Juju may hand out info with a slightly different
-	// requested volume size due to trying to adapt the unit model to the k8s world.
-	existing, err := k.getStatefulSet(spec.GetName())
-	if err != nil {
-		return errors.Trace(err)
-	}
-	// TODO(caas) - allow extra storage to be added
-	existing.Spec.Replicas = spec.Spec.Replicas
-	existing.Spec.Template.Spec.Containers = existingPodSpec.Containers
-	existing.Spec.Template.Spec.ServiceAccountName = existingPodSpec.ServiceAccountName
-	existing.Spec.Template.Spec.AutomountServiceAccountToken = existingPodSpec.AutomountServiceAccountToken
-	// NB: we can't update the Spec.ServiceName as it is immutable.
-	_, err = k.updateStatefulSet(existing)
-	return errors.Trace(err)
-}
-
-func (k *kubernetesClient) createStatefulSet(spec *apps.StatefulSet) (*apps.StatefulSet, error) {
-	out, err := k.client().AppsV1().StatefulSets(k.namespace).Create(spec)
-	if k8serrors.IsAlreadyExists(err) {
-		return nil, errors.AlreadyExistsf("stateful set %q", spec.GetName())
-	}
-	if k8serrors.IsInvalid(err) {
-		return nil, errors.NotValidf("stateful set %q", spec.GetName())
-	}
-	return out, errors.Trace(err)
-}
-
-func (k *kubernetesClient) updateStatefulSet(spec *apps.StatefulSet) (*apps.StatefulSet, error) {
-	out, err := k.client().AppsV1().StatefulSets(k.namespace).Update(spec)
-	if k8serrors.IsNotFound(err) {
-		return nil, errors.NotFoundf("stateful set %q", spec.GetName())
-	}
-	if k8serrors.IsInvalid(err) {
-		return nil, errors.NotValidf("stateful set %q", spec.GetName())
-	}
-	return out, errors.Trace(err)
-}
-
-func (k *kubernetesClient) getStatefulSet(name string) (*apps.StatefulSet, error) {
-	out, err := k.client().AppsV1().StatefulSets(k.namespace).Get(name, v1.GetOptions{IncludeUninitialized: true})
-	if k8serrors.IsNotFound(err) {
-		return nil, errors.NotFoundf("stateful set %q", name)
-	}
-	return out, errors.Trace(err)
-}
-
-// deleteStatefulSet deletes a statefulset resource.
-func (k *kubernetesClient) deleteStatefulSet(name string) error {
-	deployments := k.client().AppsV1().StatefulSets(k.namespace)
-	err := deployments.Delete(name, &v1.DeleteOptions{
-		PropagationPolicy: &defaultPropagationPolicy,
-	})
-	if k8serrors.IsNotFound(err) {
-		return nil
-	}
-
-	return errors.Trace(err)
 }
 
 func (k *kubernetesClient) deleteVolumeClaims(appName string, p *core.Pod) ([]string, error) {
